refactor(themes): remove colors from base files
