+++ src/components/avatar/themes/themes.ts
@@ -1,7 +1,5 @@
 import { css } from 'lit';
 
-// Shared Styles
-import { styles as base } from './avatar.base.css.js';
 // Dark Overrides
 import { styles as bootstrapDark } from './dark/avatar.bootstrap.css.js';
 import { styles as fluentDark } from './dark/avatar.fluent.css.js';
@@ -12,35 +10,37 @@ import { styles as bootstrapLight } from './light/avatar.bootstrap.css.js';
 import { styles as fluentLight } from './light/avatar.fluent.css.js';
 import { styles as indigoLight } from './light/avatar.indigo.css.js';
 import { styles as materialLight } from './light/avatar.material.css.js';
+// Shared Styles
+import { styles as shared } from './shared/avatar.common.css.js';
 import type { Themes } from '../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${base} ${bootstrapLight}
+    ${shared} ${bootstrapLight}
   `,
   material: css`
-    ${base} ${materialLight}
+    ${shared} ${materialLight}
   `,
   fluent: css`
-    ${base} ${fluentLight}
+    ${shared} ${fluentLight}
   `,
   indigo: css`
-    ${base} ${indigoLight}
+    ${shared} ${indigoLight}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${base} ${bootstrapDark}
+    ${shared} ${bootstrapDark}
   `,
   material: css`
-    ${base} ${materialDark}
+    ${shared} ${materialDark}
   `,
   fluent: css`
-    ${base} ${fluentDark}
+    ${shared} ${fluentDark}
   `,
   indigo: css`
-    ${base} ${indigoDark}
+    ${shared} ${indigoDark}
   `,
 };
