+++ src/components/badge/themes/shared/badge.common.scss
@@ -1,0 +1,22 @@
+@use 'styles/utilities' as *;
+@use '../light/themes' as *;
+
+$theme: $material;
+
+::slotted(igc-icon) {
+    color: var-get($theme, 'icon-color');
+}
+
+:host([variant='primary']) {
+    background: var-get($theme, 'background-color');
+    color: var-get($theme, 'text-color');
+}
+
+:host([shape='square']),
+:host([outlined][shape='square']) [part='base'] {
+    border-radius: var-get($theme, 'border-radius');
+}
+
+:host([outlined]) [part='base'] {
+    box-shadow: inset 0 0 0 rem(2px) var-get($theme, 'border-color');
+}
