+++ src/components/dialog/themes/themes.ts
@@ -6,44 +6,44 @@ import { styles as fluentDark } from './dark/dialog.fluent.css.js';
 import { styles as indigoDark } from './dark/dialog.indigo.css.js';
 import { styles as materialDark } from './dark/dialog.material.css.js';
 // Light Overrides
-import { styles as base } from './dialog.base.css.js';
 import { styles as bootstrapLight } from './light/dialog.bootstrap.css.js';
 import { styles as fluentLight } from './light/dialog.fluent.css.js';
 import { styles as indigoLight } from './light/dialog.indigo.css.js';
 import { styles as materialLight } from './light/dialog.material.css.js';
 // Shared Styles
 import { styles as bootstrap } from './shared/dialog.bootstrap.css.js';
+import { styles as shared } from './shared/dialog.common.css.js';
 import { styles as fluent } from './shared/dialog.fluent.css.js';
 import { styles as indigo } from './shared/dialog.indigo.css.js';
 import { Themes } from '../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapLight}
+    ${shared} ${bootstrap} ${bootstrapLight}
   `,
   material: css`
-    ${base} ${materialLight}
+    ${shared} ${materialLight}
   `,
   fluent: css`
-    ${fluent} ${fluentLight}
+    ${shared} ${fluent} ${fluentLight}
   `,
   indigo: css`
-    ${indigo} ${indigoLight}
+    ${shared} ${indigo} ${indigoLight}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapDark}
+    ${shared} ${bootstrap} ${bootstrapDark}
   `,
   material: css`
-    ${base} ${materialDark}
+    ${shared} ${materialDark}
   `,
   fluent: css`
-    ${fluent} ${fluentDark}
+    ${shared} ${fluent} ${fluentDark}
   `,
   indigo: css`
-    ${indigo} ${indigoDark}
+    ${shared} ${indigo} ${indigoDark}
   `,
 };
