+++ src/components/expansion-panel/themes/expansion-panel.base.scss
@@ -1,8 +1,5 @@
 @use 'styles/common/component';
 @use 'styles/utilities' as *;
-@use './light/themes' as *;
-
-$theme: $material;
 
 :host {
     --lines-clamped: 1;
@@ -11,9 +8,6 @@ $theme: $material;
     overflow: hidden;
     display: flex;
     flex-direction: column;
-    color: var-get($theme, 'body-color');
-    background: var-get($theme, 'body-background');
-    border-radius: var-get($theme, 'border-radius');
 }
 
 [part='header'] {
@@ -22,12 +16,10 @@ $theme: $material;
     cursor: pointer;
     padding-inline: pad-inline(24px);
     padding-block: pad-block(16px);
-    background: var-get($theme, 'header-background');
 
     &:focus,
     &:active {
         outline: transparent;
-        background: var-get($theme, 'header-focus-background');
     }
 
     *:not([part~='indicator']) {
@@ -40,20 +32,14 @@ $theme: $material;
     }
 }
 
-[part='title'] {
-    color: var-get($theme, 'header-title-color');
+[part='title']::slotted(*) {
+    @include type-style('h5');
+    @include line-clamp(var(--lines-clamped), true, true);
 
-    &::slotted(*) {
-        @include type-style('h5');
-        @include line-clamp(var(--lines-clamped), true, true);
-
-        margin: 0;
-    }
+    margin: 0;
 }
 
 [part='subtitle'] {
-    color: var-get($theme, 'header-description-color');
-
     &::slotted(*) {
         @include type-style('subtitle-2');
         @include ellipsis();
@@ -83,8 +69,6 @@ $theme: $material;
 }
 
 [part~='indicator'] {
-    color: var-get($theme, 'header-icon-color');
-
     igc-icon {
         --ig-size: 3;
     }
@@ -105,10 +89,4 @@ $theme: $material;
 
 :host([disabled]) {
     pointer-events: none;
-
-    [part='title'],
-    [part='subtitle'],
-    [part~='indicator'] {
-        color: var-get($theme, 'disabled-color');
-    }
 }
