+++ src/components/expansion-panel/themes/shared/expansion-panel.common.scss
@@ -1,0 +1,39 @@
+@use 'styles/utilities' as *;
+@use '../light/themes' as *;
+
+$theme: $material;
+
+:host {
+    color: var-get($theme, 'body-color');
+    background: var-get($theme, 'body-background');
+    border-radius: var-get($theme, 'border-radius');
+}
+
+[part='header'] {
+    background: var-get($theme, 'header-background');
+
+    &:focus,
+    &:active {
+        background: var-get($theme, 'header-focus-background');
+    }
+}
+
+[part='title'] {
+    color: var-get($theme, 'header-title-color');
+}
+
+[part='subtitle'] {
+    color: var-get($theme, 'header-description-color');
+}
+
+[part~='indicator'] {
+    color: var-get($theme, 'header-icon-color');
+}
+
+:host([disabled]) {
+    [part='title'],
+    [part='subtitle'],
+    [part~='indicator'] {
+        color: var-get($theme, 'disabled-color');
+    }
+}
