+++ src/components/snackbar/themes/themes.ts
@@ -12,38 +12,38 @@ import { styles as indigoLight } from './light/snackbar.indigo.css.js';
 import { styles as materialLight } from './light/snackbar.material.css.js';
 // Shared Styles
 import { styles as bootstrap } from './shared/snackbar.bootstrap.css.js';
+import { styles as shared } from './shared/snackbar.common.css.js';
 import { styles as fluent } from './shared/snackbar.fluent.css.js';
 import { styles as indigo } from './shared/snackbar.indigo.css.js';
-import { styles as base } from './snackbar.base.css.js';
 import { Themes } from '../../../theming/types.js';
 
 const light = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapLight}
+    ${shared} ${bootstrap} ${bootstrapLight}
   `,
   material: css`
-    ${base} ${materialLight}
+    ${shared} ${materialLight}
   `,
   fluent: css`
-    ${fluent} ${fluentLight}
+    ${shared} ${fluent} ${fluentLight}
   `,
   indigo: css`
-    ${indigo} ${indigoLight}
+    ${shared} ${indigo} ${indigoLight}
   `,
 };
 
 const dark = {
   bootstrap: css`
-    ${bootstrap} ${bootstrapDark}
+    ${shared} ${bootstrap} ${bootstrapDark}
   `,
   material: css`
-    ${base} ${materialDark}
+    ${shared} ${materialDark}
   `,
   fluent: css`
-    ${fluent} ${fluentDark}
+    ${shared} ${fluent} ${fluentDark}
   `,
   indigo: css`
-    ${indigo} ${indigoDark}
+    ${shared} ${indigo} ${indigoDark}
   `,
 };
