+++ src/components/toast/themes/shared/toast.common.scss
@@ -1,0 +1,13 @@
+@use 'styles/utilities' as *;
+@use '../light/themes' as *;
+
+$theme: $material;
+
+:host {
+    padding: rem(10px) rem(16px);
+    color: var-get($theme, 'text-color');
+    background: var-get($theme, 'background');
+    border: #{rem(1px)} solid var-get($theme, 'border-color');
+    box-shadow: var-get($theme, 'elevation');
+    border-radius: var-get($theme, 'border-radius');
+}
